feat: 예약과 관련된 테마 삭제시 예외 발생 기능 추가

+++ src/main/java/nextstep/main/java/nextstep/service/ThemeService.java
@@ -4,7 +4,8 @@ import nextstep.main.java.nextstep.domain.Theme;
 import nextstep.main.java.nextstep.domain.ThemeCreateRequestDto;
 import nextstep.main.java.nextstep.exception.exception.DuplicateThemeException;
 import nextstep.main.java.nextstep.exception.exception.NoSuchThemeException;
-import nextstep.main.java.nextstep.message.ExceptionMessage;
+import nextstep.main.java.nextstep.exception.exception.RelatedThemeDeletionException;
+import nextstep.main.java.nextstep.repository.ReservationRepository;
 import nextstep.main.java.nextstep.repository.ThemeRepository;
 import org.springframework.stereotype.Service;
 
@@ -16,9 +17,11 @@ import static nextstep.main.java.nextstep.message.ExceptionMessage.*;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ThemeService(ThemeRepository themeRepository) {
+    public ThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public Theme createTheme(ThemeCreateRequestDto themeCreateRequestDto) {
@@ -37,6 +40,10 @@ public class ThemeService {
         if (themeRepository.findById(id)
                 .isPresent()) {
             throw new NoSuchThemeException(NO_SUCH_THEME_MESSAGE);
+        }
+        if (!reservationRepository.findAllByThemeId(id)
+                .isEmpty()) {
+            throw new RelatedThemeDeletionException(RELATED_THEME_DELETION_MESSAGE);
         }
         themeRepository.deleteById(id);
     }
